refactor: Move iProd functions in Toolkit back to iProd Class
