+++ lib/Toolkit.js
@@ -1376,375 +1376,7 @@ class Toolkit {
     this.conn.run(toJson); // Post the input XML and get the response.
   }
 
-  getPTFInfo(PTFID, cb) {
-    const outBuf = {
-      type: 'ds',
-      io: 'out',
-      len: 'rec1',
-      fields: [
-        { type: '10i0', value: 0 }, // Bytes returned
-        { type: '10i0', value: 0 }, // Bytes available
-        { type: '10i0', value: 0 }, // Offset to additional information
-        { type: '7A', value: '' }, // Product ID
-        { type: '7A', value: '' }, // PTF ID
-        { type: '6A', value: '' }, // Release level
-        { type: '4A', value: '' }, // Product option
-        { type: '4A', value: '' }, // Load ID
-        { type: '1A', value: '' }, // Loaded status
-        { type: '1A', value: '' }, // Cover letter status
-        { type: '1A', value: '' }, // On-order status
-        { type: '1A', value: '' }, // Save file status
-        { type: '10A', value: '' }, // File name
-        { type: '10A', value: '' }, // File library name
-        { type: '1A', value: '' }, // PTF type
-        { type: '1A', value: '' }, // IPL action
-        { type: '1A', value: '' }, // Action pending
-        { type: '1A', value: '' }, // Action required
-        { type: '1A', value: '' }, // PTF is released
-        { type: '6A', value: '' }, // Target release
-        { type: '7A', value: '' }, // Superseding PTF
-        { type: '1A', value: '' }, // Current IPL source
-        { type: '2A', value: '' }, // Minimum level
-        { type: '2A', value: '' }, // Maximum level
-        { type: '1A', value: '' }, // Format information available
-        { type: '13A', value: '' }, // Status date and time
-        { type: '7A', value: '' }, // Licensed Internal Code group
-        { type: '7A', value: '' }, // Superseded by PTF ID
-        { type: '1A', value: '' }, // Current server IPL source
-        { type: '1A', value: '' }, // Server IPL required
-        { type: '13A', value: '' }, // Creation date and time
-        { type: '1A', value: '' }, // Technology refresh PTF
-        { type: '1A', value: '' }, // Reserved
-      ],
-    };
 
-    const PTFInfo = {
-      type: 'ds',
-      fields: [
-        { type: '7A', value: PTFID }, // PTF ID
-        { type: '7A', value: '*ONLY' }, // Product ID
-        { type: '6A', value: '' }, // Release level
-        { type: '10i0', value: '' }, // CCSID for returned directory names
-        { type: '1A', value: '0' }, // Close PTF database files
-        { type: '25A', value: '' }, // Reserved
-      ],
-    };
-
-    const pgm = new ProgramCall('QPZRTVFX', { lib: 'QSYS' });
-    pgm.addParam(outBuf);
-    pgm.addParam({ type: '10i0', setlen: 'rec1', value: 0 });
-    pgm.addParam(PTFInfo);
-    pgm.addParam({ type: '8A', value: 'PTFR0100' });
-    pgm.addParam(this.errno);
-
-    this.conn.add(pgm.toXML());
-
-    let rtValue; // The returned value.
-
-    const toJson = (transportError, str) => {
-      if (transportError) {
-        if (this.reportError) {
-          cb(transportError, null);
-          return;
-        }
-        cb(null);
-        return;
-      }
-      parseString(str, (parseError, result) => {
-        if (parseError) {
-          if (this.reportError) {
-            cb(parseError, null);
-            return;
-          }
-          cb(null);
-          return;
-        }
-        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes('+++ success')) {
-          const { data } = result.myscript.pgm[0].parm[0].ds[0];
-          rtValue = {
-            Product_ID: data[3]._,
-            PTF_ID: data[4]._,
-            Release_level: data[5]._,
-            Product_option: data[6]._,
-            Load_ID: data[7]._,
-            Loaded_status: data[8]._,
-            Cover_letter_status: data[9]._,
-            'On-order_status': data[10]._,
-            Save_file_status: data[11]._,
-            File_name: data[12]._,
-            File_library_name: data[13]._,
-            PTF_type: data[14]._,
-            IPL_action: data[15]._,
-            Action_pending: data[16]._,
-            Action_required: data[17]._,
-            PTF_is_released: data[18]._,
-            Target_release: data[19]._,
-            Superseding_PTF: data[20]._,
-            Current_IPL_source: data[21]._,
-            Minimum_level: data[22]._,
-            Maximum_level: data[23]._,
-            Format_information_available: data[24]._,
-            Status_date_and_time: data[25]._,
-            Licensed_Internal_Code_group: data[26]._,
-            Superseded_by_PTF_ID: data[27]._,
-            Current_server_IPL_source: data[28]._,
-            Server_IPL_required: data[29]._,
-            Creation_date_and_time: data[30]._,
-            Technology_refresh_PTF: data[31]._,
-            Reserved: data[32]._,
-          };
-        } else { rtValue = str; }
-
-        if (this.reportError) {
-          cb(null, rtValue);
-          return;
-        }
-        cb(rtValue);
-      });
-    };
-
-    this.conn.run(toJson); // Post the input XML and get the response.
-  }
-
-  getProductInfo(prodID, option, cb) {
-    let callback = cb;
-    let prodOption;
-    const outBuf = {
-      type: 'ds',
-      io: 'out',
-      len: 'rec1',
-      fields: [
-        { value: 0, type: '10i0' }, // Bytes returned
-        { value: 0, type: '10i0' }, // Bytes available
-        { value: 0, type: '10i0' }, // Reserved
-        { type: '7A', value: '' }, // Product ID
-        { type: '6A', value: '' }, // Release level
-        { type: '4A', value: '' }, // Product option
-        { type: '4A', value: '' }, // Load ID
-        { type: '10A', value: '' }, // Load type
-        { type: '10A', value: '' }, // Symbolic load state
-        { type: '10A', value: '' }, // Load error indicator
-        { type: '2A', value: '' }, // Load state
-        { type: '1A', value: '' }, // Supported flag
-        { type: '2A', value: '' }, // Registration type
-        { type: '14A', value: '' }, // Registration value
-        { type: '2A', value: '' }, // Reserved
-        { type: '10i0', value: 0 }, // Offset to additional information
-        { type: '4A', value: '' }, // Primary language load identifier
-        { type: '6A', value: '' }, // Minimum target release
-        { type: '6A', value: '' }, // Minimum VRM of *BASE required by option
-        { type: '1A', value: '' }, // Requirements met between base and option value
-        { type: '3A', value: '' }, // Level
-        { type: '1A', value: '' }, // Reserved
-      ],
-    };
-    if (!cb) { // If we do not get the third param,
-      if (option) { // If we get two params,
-        if (typeof option === 'function') { // If it is a function,
-          callback = option; // then it is the callback.
-          prodOption = '0000';
-        }
-      } else { // If we have only one param,
-        prodOption = '0000'; // then use *BASE as default.
-      }
-    }
-
-    if (Number.isInteger(option)) {
-      if (option < 0 || option > 99) {
-        prodOption = '0000';
-      } else if (prodOption < 10) {
-        prodOption = `000${option}`;
-      } else {
-        prodOption = `00${option}`;
-      }
-    }
-
-    const ProdInfo = {
-      type: 'ds',
-      fields: [
-        { type: '7A', value: prodID },
-        { type: '6A', value: '*ONLY' },
-        { type: '4A', value: prodOption },
-        { type: '10A', value: '*CODE' },
-      ],
-    };
-
-    const pgm = new ProgramCall('QSZRTVPR', { lib: 'QSYS' });
-    pgm.addParam(outBuf);
-    pgm.addParam({ type: '10i0', setlen: 'rec1', value: 0 });
-    pgm.addParam({ type: '8A', value: 'PRDR0100' });
-    pgm.addParam(ProdInfo);
-    pgm.addParam(this.errno);
-
-    this.conn.add(pgm.toXML());
-
-    let rtValue; // The returned value.
-    const toJson = (transportError, str) => {
-      if (transportError) {
-        if (this.reportError) {
-          callback(transportError, null);
-          return;
-        }
-        callback(null);
-        return;
-      }
-      parseString(str, (parseError, result) => {
-        if (parseError) {
-          if (this.reportError) {
-            callback(parseError, null);
-            return;
-          }
-          callback(null);
-          return;
-        }
-        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes('+++ success')) {
-          const { data } = result.myscript.pgm[0].parm[0].ds[0];
-
-          rtValue = {
-            // Reserved: data[2]._, // TODO: Duplicate key
-            Product_ID: data[3]._,
-            Release_level: data[4]._,
-            Product_option: data[5]._,
-            Load_ID: data[6]._,
-            Load_type: data[7]._,
-            Symbolic_load_state: data[8]._,
-            Load_error_indicator: data[9]._,
-            Load_state: data[10]._,
-            Supported_flag: data[11]._,
-            Registration_type: data[12]._,
-            Registration_value: data[13]._,
-            // Reserved: data[14]._, // TODO: Duplicate key
-            Offset_to_additional_information: data[15]._,
-            Primary_language_load_identifier: data[16]._,
-            Minimum_target_release: data[17]._,
-            'Minimum_VRM_of_*BASE_required_by_option': data[18]._,
-            Requirements_met_between_base_and_option_value: data[19]._,
-            Level: data[20]._,
-            Reserved: data[21]._,
-          };
-        } else { rtValue = str; }
-
-        if (this.reportError) {
-          callback(null, rtValue);
-          return;
-        }
-        callback(rtValue);
-      });
-    };
-
-    this.conn.run(toJson); // Post the input XML and get the response.
-  }
-
-  getInstalledProducts(cb) {
-    const maxProd = 197;
-    const outBuf = {
-      type: 'ds',
-      io: 'out',
-      fields: [],
-    };
-    for (let i = 0; i < maxProd; i += 1) {
-      outBuf.fields.push({ type: '7A', value: '' }); // Product ID
-      outBuf.fields.push({ type: '5A', value: '' }); // Product option
-      outBuf.fields.push({ type: '6A', value: '' }); // Release level
-      outBuf.fields.push({ type: '2h', value: '' }); // Skip
-      outBuf.fields.push({ type: '7A', value: '' }); // Description text message ID
-      outBuf.fields.push({ type: '10A', value: '' }); // Description text object name
-      outBuf.fields.push({ type: '10A', value: '' }); // Description text library name
-      outBuf.fields.push({ type: '1A', value: '' }); // Installed flag
-      outBuf.fields.push({ type: '1A', value: '' }); // Supported flag
-      outBuf.fields.push({ type: '2A', value: '' }); // Registration type
-      outBuf.fields.push({ type: '14A', value: '' }); // Registration value
-      outBuf.fields.push({ type: '132A', value: '' }); // Description text
-    }
-    const inputInfo = {
-      type: 'ds',
-      fields: [
-        { type: '10i0', value: maxProd },
-        { type: '10A', value: '*ALL' },
-        { value: '1', type: '1A' },
-        { value: '1', type: '1A' },
-        { type: '10A', value: '*ALL' },
-        { type: '10A', value: '*INSTLD' },
-        { type: '10i0', value: maxProd },
-      ],
-    };
-    const ProdInfo = {
-      type: 'ds',
-      fields: [
-        { type: '7A', value: '' },
-        { type: '5A', value: '' },
-        { type: '6A', value: '' },
-      ],
-    };
-    const OutputInfo = {
-      type: 'ds',
-      io: 'out',
-      fields: [
-        { type: '10i0', value: 0 },
-        { type: '10i0', value: 0 },
-        { type: '10i0', value: 0 },
-      ],
-    };
-    const pgm = new ProgramCall('QSZSLTPR', { lib: 'QSYS' });
-    pgm.addParam(outBuf);
-    pgm.addParam(inputInfo);
-    pgm.addParam({ type: '8A', value: 'PRDS0200' });
-    pgm.addParam(ProdInfo);
-    pgm.addParam(OutputInfo);
-    pgm.addParam(this.errno);
-
-    this.conn.add(pgm.toXML());
-
-    let rtValue = []; // The returned value.
-    const toJson = (transportError, str) => {
-      if (transportError) {
-        if (this.reportError) {
-          cb(transportError, null);
-          return;
-        }
-        cb(null);
-        return;
-      }
-      parseString(str, (parseError, result) => {
-        if (parseError) {
-          if (this.reportError) {
-            cb(parseError, null);
-            return;
-          }
-          cb(str);
-          return;
-        }
-        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes('+++ success')) {
-          const { data } = result.myscript.pgm[0].parm[0].ds[0];
-          const count = result.myscript.pgm[0].parm[4].ds[0].data[1]._;
-
-          for (let i = 0; i < count * 12; i += 12) {
-            rtValue.push({
-              Product_ID: data[i]._,
-              Product_option: data[i + 1]._,
-              Release_level: data[i + 2]._,
-              Description_text_message_ID: data[i + 4]._,
-              Description_text_object_name: data[i + 5]._,
-              Description_text_library_name: data[i + 6]._,
-              Installed_flag: data[i + 7]._,
-              Supported_flag: data[i + 8]._,
-              Registration_type: data[i + 9]._,
-              Registration_value: data[i + 10]._,
-              Description_text: data[i + 11]._,
-            });
-          }
-        } else { rtValue = str; }
-
-        if (this.reportError) {
-          cb(null, rtValue);
-          return;
-        }
-        cb(rtValue);
-      });
-    };
-
-    this.conn.run(toJson); // Post the input XML and get the response.
-  }
 
 
   createUserSpace(name, lib, attr, size, auth, desc, cb) {
